+++ attributes/default.rb
@@ -25,4 +25,4 @@
 # OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
 # SUCH DAMAGE.
 
-default['tinyos']['main-dir'] = "/usr/src/tinyos-main"
+default['chef-tinyos']['main-dir'] = "/usr/src/tinyos-main"