+++ recipes/install_avr_dev.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: tinyos
+# Cookbook Name:: chef-tinyos
 # Recipe:: install_avr_dev
 #
 # Copyright (c) 2014, Shigemi ISHIDA