+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-tinyos CHANGELOG
+chef-tinyos CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the tinyos cookbook.
+This file is used to list changes made in each version of the chef-tinyos cookbook.
 
 0.1.0
 -----
-- [Shigemi ISHIDA] - Initial release of tinyos
+- [Shigemi ISHIDA] - Initial release of chef-tinyos